upload file, delete:comm & post

+++ first_app/api_app/comment_api.py
@@ -37,37 +37,71 @@ def create_comment(post_id):
     return jsonify(comment_schema.dump(new_comm))
 
 
-@comment_urls.route('/api/v1/<comment_id>/update-comments/<user_id>', methods=['PUT']) 
+@comment_urls.route('/api/v1/update-comments/<int:comment_id>', methods=['PUT']) 
 @login_required   
-def update_comment(comment_id, user_id):
+def update_comment(comment_id):
     """Редагування(оновлення) коментарів"""
     data = request.json
     text = data['text']
-    created = data['created']
+    created = datetime.datetime.utcnow()
 
     comment = Comments.query.filter(
         Comments.id == comment_id,
-        Comments.author_id == user_id
+        Comments.author_id == g.user_id
     ).first()
+
     if not comment:
         return {"Error": "This user hasn't comments"}
     
     comment.text = text
-    comment.created = datetime.datetime.now()
+    comment.created = datetime.datetime.utcnow()
 
     db.session.commit()
 
-    comment_schema = CommentsSchema(many=True)
+    comment_schema = CommentsSchema()
 
     return jsonify(comment_schema.dump(comment))
 
 
-@comment_urls.route('/api/v1/<user_id>/delete-comment', methods=['DELETE'])
+@comment_urls.route('/api/v1/delete-comments/<int:comment_id>', methods=['DELETE'])
 @login_required
-def delete_comment(user_id):
+def delete_comments(comment_id):
     """Видалення коментарів"""
-    comment = Comments.query.get(user_id)
+    author_id = g.user_id
+    comment = Comments.query.filter(Comments.post_id==comment_id).all()
+
+    if comment or not comment:
+        # return {"Error": "Comment not found"}, 400 
+        for comments in comment:
+            db.session.delete(comments)
+            db.session.commit()
+
+    return {"Status": "Comments deleted"}, 200
+
+
+@comment_urls.route('/api/v1/delete-comment/<int:comment_id>', methods=['DELETE'])
+@login_required
+def delete_one_comment(comment_id):
+    """Видалення одного коментаря"""
+    author_id = g.user_id
+    comment = Comments.query.filter(Comments.id==comment_id).first()
+
+    if not comment:
+        return {"Error": "Comment not found"}, 400
+     
     db.session.delete(comment)
     db.session.commit()
 
-    return {"Status": "Comment deleted"}
+    return {"Status": "One Comment deleted"}, 200
+
+
+@comment_urls.route('/api/v1/comment/<int:comment_id>', methods=['GET'])
+@login_required
+def user_comment(comment_id):
+    """Виводимо коментар користувача"""
+
+    author_id = g.user_id
+    comment = Comments.query.filter(Comments.id == comment_id).first()
+    comment_schema = CommentsSchema()
+
+    return jsonify(comment_schema.dump(comment))
